+++ src/test/java/roomescape/business/service/ReservationTimeServiceTest.java
@@ -1,4 +1,4 @@
-package roomescape.service;
+package roomescape.business.service;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
@@ -11,7 +11,6 @@ import org.mockito.Mockito;
 import roomescape.business.model.entity.ReservationTime;
 import roomescape.business.model.repository.ReservationDao;
 import roomescape.business.model.repository.ReservationTimeDao;
-import roomescape.business.service.ReservationTimeService;
 import roomescape.presentation.dto.request.ReservationTimeRequest;
 import roomescape.presentation.dto.response.ReservationTimeResponse;
