+++ src/test/java/roomescape/business/service/ThemeServiceTest.java
@@ -1,4 +1,4 @@
-package roomescape.service;
+package roomescape.business.service;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
@@ -10,7 +10,6 @@ import org.mockito.Mockito;
 import roomescape.business.model.entity.Theme;
 import roomescape.business.model.repository.ReservationDao;
 import roomescape.business.model.repository.ThemeDao;
-import roomescape.business.service.ThemeService;
 import roomescape.presentation.dto.request.ThemeRequest;
 import roomescape.presentation.dto.response.ThemeResponse;
