+++ src/test/java/roomescape/infrastructure/ReservationDaoImplTest.java
@@ -1,4 +1,4 @@
-package roomescape.repository;
+package roomescape.infrastructure;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
@@ -19,7 +19,6 @@ import roomescape.business.model.entity.Reservation;
 import roomescape.business.model.entity.ReservationTime;
 import roomescape.business.model.entity.Theme;
 import roomescape.business.model.repository.ReservationDao;
-import roomescape.infrastructure.ReservationDaoImpl;
 
 @JdbcTest
 @ActiveProfiles("test")
